requirements.txt
# Core dependencies
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
PyYAML==6.0.1

# Feature engineering and preprocessing
category_encoders==2.6.1
feature-engine==1.6.1

# Model training and evaluation
xgboost==1.7.6
lightgbm==4.0.0
shap==0.42.1
optuna==3.3.0

# Utilities
tqdm==4.65.0
joblib==1.3.1
pytest==7.4.0
python-dotenv==1.0.0

# Logging
loguru==0.7.0
